Add test environment validation

+++ conftest.py
@@ -5,6 +5,10 @@ from app import db, app
 
 @pytest.fixture
 def client():
+    if app.config['ENV'] != 'testing':
+        raise EnvironmentError(
+            "It's not testing environment! Set correct environment variable."
+        )
     db.create_all()
     yield app.test_client()
     db.drop_all()
